test(ui): move the timer-chain assertion beside the timer-chain code
